+++ core/src/components/popover/popover.tsx
@@ -1,11 +1,8 @@
 import type { ComponentInterface, EventEmitter } from '@stencil/core';
 import { Component, Element, Event, Host, Method, Prop, State, Watch, h } from '@stencil/core';
-
-import { getIonMode } from '../../global/ionic-global';
-import type { AnimationBuilder, ComponentProps, ComponentRef, FrameworkDelegate } from '../../interface';
-import { CoreDelegate, attachComponent, detachComponent } from '../../utils/framework-delegate';
-import { addEventListener, raf, hasLazyBuild } from '../../utils/helpers';
-import { printIonWarning } from '../../utils/logging';
+import { CoreDelegate, attachComponent, detachComponent } from '@utils/framework-delegate';
+import { addEventListener, raf, hasLazyBuild } from '@utils/helpers';
+import { printIonWarning } from '@utils/logging';
 import {
   BACKDROP,
   dismiss,
@@ -14,11 +11,14 @@ import {
   prepareOverlay,
   present,
   setOverlayId,
-} from '../../utils/overlays';
+} from '@utils/overlays';
+import { isPlatform } from '@utils/platform';
+import { getClassMap } from '@utils/theme';
+import { deepReady, waitForMount } from '@utils/transition';
+
+import { getIonMode } from '../../global/ionic-global';
+import type { AnimationBuilder, ComponentProps, ComponentRef, FrameworkDelegate } from '../../interface';
 import type { OverlayEventDetail } from '../../utils/overlays-interface';
-import { isPlatform } from '../../utils/platform';
-import { getClassMap } from '../../utils/theme';
-import { deepReady, waitForMount } from '../../utils/transition';
 
 import { iosEnterAnimation } from './animations/ios.enter';
 import { iosLeaveAnimation } from './animations/ios.leave';